analytics/requirements.txt
paho-mqtt==2.1.0
httpx==0.27.2
nats-py==2.9.0